pyproject.toml
[tool.poetry]
name = "crescendo-backend"
version = "0.1.0"
description = ""
authors = ["tgoddessana <twicegoddessana1229@gmail.com>"]
readme = "README.md"
packages = [{include = "crescendo_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
fullask-rest-framework = "*"
google-auth = "^2.19.1"
requests = "^2.31.0"
gunicorn = "^20.1.0"
oracledb = "^1.3.2"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
pytest = "^7.3.2"
flake8 = "^6.0.0"
pre-commit = "^3.3.2"
flask-moreshell = "^0.0.2"
ipython = "^8.14.0"
